% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R, R/SaemixData.R
\docType{methods}
\name{show-methods}
\alias{show-methods}
\alias{show,SaemixData-method}
\alias{show,SaemixRepData-method}
\alias{show,SaemixSimData-method}
\title{Methods for Function show}
\usage{
\S4method{show}{SaemixData}(object)

\S4method{show}{SaemixRepData}(object)

\S4method{show}{SaemixSimData}(object)
}
\description{
Prints a short summary of an object
}
\section{Methods}{
 \describe{

\item{list("signature(x = \"ANY\")")}{ Default show function }

\item{list("signature(x = \"SaemixData\")")}{ Prints a short summary of a
SaemixData object }

\item{list("signature(x = \"SaemixModel\")")}{ Prints a short summary of a
SaemixModel object }

\item{list("signature(x = \"SaemixObject\")")}{ Prints a short summary of
the results from a SAEMIX fit }

\item{list("signature(x = \"SaemixRes\")")}{ Not user-level }

\item{list("signature(object = \"SaemixRepData\")")}{ Prints a short summary
of a SaemixRepData object }

\item{list("signature(object = \"SaemixSimData\")")}{ Prints a short summary
of a SaemixSimData object } }
}

\keyword{methods}
