% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R
\docType{methods}
\name{summary-methods}
\alias{summary-methods}
\alias{summary,ANY-method}
\title{Methods for Function summary}
\description{
Methods for function \code{summary}
}
\section{Methods}{
 \describe{

\item{list("signature(x = \"ANY\")")}{ default summary function ?}

\item{list("signature(x = \"SaemixData\")")}{ summary of the data }

\item{list("signature(x = \"SaemixModel\")")}{ summary of the model }

\item{list("signature(x = \"SaemixObject\")")}{ summary of an SaemixObject}

}
}

\keyword{methods}
