% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixData.R
\docType{methods}
\name{summary,SaemixData-method}
\alias{summary,SaemixData-method}
\alias{summary}
\alias{summary,SaemixData}
\title{Summarising longitudinal data}
\usage{
\S4method{summary}{SaemixData}(object, print = TRUE, ...)
}
\arguments{
\item{object}{an object of clas SaemixData}

\item{print}{a boolean controlling whether to print the output or return it silently}

\item{...}{additional arguments (ignored)}
}
\value{
a list with a number of elements extracted from the dataset
\describe{
\item{N}{ number of subjects}
\item{nobs}{ the total number of observations} 
\item{nind.obs}{a vector giving the number of observations for each subject}
\item{id}{subject ID; x: predictors; y: response, and, if present in the data, covariates: the covariates (as many lines as observations) and ind.covariates: the individual covariates (one}
}
}
\description{
summary method for class SaemixData
}
