% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixData.R
\name{transformCatCov}
\alias{transformCatCov}
\title{Transform covariates}
\usage{
transformCatCov(object, covariate, group, reference, verbose = FALSE)
}
\arguments{
\item{object}{saemixData object}

\item{covariate}{name of the covariate}

\item{group}{a vector giving the categories to which the initial values of the covariates should be mapped. If the resulting covariate is binary, it will be stored as 0/1. If it has more than 2 categories, dummy covariates will be created for the analysis.}

\item{reference}{the reference group}

\item{verbose}{a boolean, prints messages during the execution of the function if TRUE. Defaults to FALSE.}
}
\value{
an object of class \code{"\linkS4class{SaemixData}"}
}
\description{
Regroup categorical covariates
}
\examples{
# TODO
}
\keyword{data}
